% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Maps all the incidents occuring in a given state for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numeric identifier of the state that is of interest.}

\item{year}{The integer year that is of interest.}
}
\value{
A graphic showing a map of the state of interest with the FARS incidents overlaid.
}
\description{
The FARS is a nationwide census prepared by the US National Highway Traffic Safety Administration.
It provides the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\details{
Years must be represented by numeric values. Years represented as character strings will cause an error.

FARS files should be situated in the working directory.
FARS files can be downloaded from the NHTSA website - \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}.

Use of an invalid state number will cause an error.
The list of state numbers can be found in FARS user manual - \url{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812602}
}
\examples{
\dontrun{fars_map_state(1, 2013)}

}
