% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a Fatality Analysis Reporting System (FARS) file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that represents the location of the FAR file on your computer.}
}
\value{
A Tibble containing the information in the FARS file.
}
\description{
The FARS is a nationwide census prepared by the US National Highway Traffic Safety Administration.
It provides the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\details{
FARS files can be downloaded from the NHTSA website - \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}.

This function reads a FARS file into the R namespace.
}
\examples{
\dontrun{fars_read("./data/accident_2013.csv.bz2")}

}
