% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads multiple Fatality Analysis Reporting System (FARS) files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of integer years that are of interest.}
}
\value{
A Tibble containing the information in a FARS file for each year of interest.
}
\description{
The FARS is a nationwide census prepared by the US National Highway Traffic Safety Administration.
It provides the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\details{
Years must be represented by numeric values. Years represented as character strings will cause an error.

FARS files should be situated in the working directory.
FARS files can be downloaded from the NHTSA website - \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}.
}
\examples{
\dontrun{fars_read_years(2013:2015)}

}
