% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{The information in FARS files for a given set of years is aggregated to provide a count of the number of incidents occuring each month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of integer years that are of interest.}
}
\value{
A Tibble containing the summarised information. The Tibble will have three columns - year, MONTH, and n.
}
\description{
The FARS is a nationwide census prepared by the US National Highway Traffic Safety Administration.
It provides the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\details{
Years must be represented by numeric values. Years represented as character strings will cause an error.

FARS files should be situated in the working directory.
FARS files can be downloaded from the NHTSA website - \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}.
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}

}
